+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Added
 
 - Added `navbar.jsx` with a logo, navigation menu, and sign-up button.
+- Modularized `GetData()` function
 
 ## [0.1.11] - 2025-04-18